# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VEP C2M Service API",
        "description": "VEP C2M Service API",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "",
            "description": "clouddev"
        },
        {
            "url": "",
            "description": "sit"
        },
        {
            "url": "",
            "description": "uat"
        },
        {
            "url": "",
            "description": "prd"
        }
    ],
    "components": {
        "schemas": {
...
            "ConnectionHandleDto": {
                "type": "object",
                "properties": {}
            },
            "ConnectionTargetActionDto": {
                "type": "object",
                "properties": {}
            },
            "ConnectionRenameActionDto": {
                "type": "object",
                "properties": {}
            },
...
            "ConnectionActionDto": {
                "type": "object",
                "properties": {}
            },
...
            "DisconnectHandleDto": {
                "type": "object",
                "properties": {}
            },
...
        }
    },
    "paths": {
...
        "/cbm/fairs/{fairCode}/meetings/{id}/video-conference": {
            "post": {
                "operationId": "CBMVideoConferenceController_startVideoConference",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "CBMVideoConferenceController_endVideoConfernece",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/cbm/fairs/{fairCode}/meetings/{id}/video-conference/switch-to-zoom": {
            "post": {
                "operationId": "CBMVideoConferenceController_switchToZoom",
                "parameters": [
                    {
                        "name": "fairCode",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cbm/video-conference/join": {
            "post": {
                "operationId": "CBMVideoConferenceController_wsJoinRoom",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnectionHandleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cbm/video-conference/mute": {
            "post": {
                "operationId": "CBMVideoConferenceController_wsMutePeople",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnectionTargetActionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cbm/video-conference/unmute": {
            "post": {
                "operationId": "CBMVideoConferenceController_wsUnmutePeople",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnectionTargetActionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cbm/video-conference/kick": {
            "post": {
                "operationId": "CBMVideoConferenceController_wsKickPeople",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnectionTargetActionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cbm/video-conference/admit": {
            "post": {
                "operationId": "CBMVideoConferenceController_wsAdmitPeople",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnectionTargetActionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cbm/video-conference/reject": {
            "post": {
                "operationId": "CBMVideoConferenceController_wsRejectPeople",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnectionTargetActionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cbm/video-conference/rename": {
            "post": {
                "operationId": "CBMVideoConferenceController_wsRename",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnectionRenameActionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cbm/video-conference/confirm-extend-meeting": {
            "post": {
                "operationId": "CBMVideoConferenceController_wsConfirmExtendMeeting",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConnectionActionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cbm/video-conference/disconnect": {
            "post": {
                "operationId": "CBMVideoConferenceController_wsDisconnect",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DisconnectHandleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
    }
}